contrib: Add a place for helpful files
The files in `contrib/` are intended to be examples helpful
illustrations when installing, configuring, and using
`soundcraft-utils`, mainly by giving some ideas which can help
you find a solution.

These files are not intended to be working out of the box, be
documented flawlessly, or be useful for everybody.

If one of these files directly works for you unmodified, that
is a lucky coincidence. Otherwise, use modified copies of them
or use them as inspiration for your own solutions.