contrib: Command line DBus example set MASTER_L_R
In the command line DBus example, set the USB audio routing to
MASTER_L_R as that is be available on all three Notepad devices.

Now the example script should run out of the box with any
Notepad model, not just on the Notepad-12FX which the script's
author happens to own.