contrib: Add todo.sh to list TODO and FIXME from the sources
Add one script to find all TODOs and FIXMEs in the source tree.

This has a bit of an issue with multi-line TODOs and FIXMEs,
but it is a good start.

Example output:

```
[user@host soundcraft-utils]$ ./contrib/sources/todo.sh
README.md:TODO
soundcraft/gui.py:            # Todo: Can we relaunch ourselves?
soundcraft/notepad.py:        # TODO: Decode these?
[user@host soundcraft-utils]$
```