gui: allow multiple instances of the GUI app
Allow multiple instances of the GUI app to run at the same time.

This can be useful to

   * test the communication via the D-Bus service by having two
     GUI windows side by side (for developers)

   * have a GUI window on more than one virtual desktop (for users,
     like other audio settings programs like pavucontrol)

However, some programs have started to be single GUI window only
which forces virtual desktop switches, but that might be a new
Gnome desktop policy other just changed default behaviour in Gtk
without much more background.

Note that `flake8` and `black` disagree about how the expression
ORing the binary constants is supposed to be formatted, so we
resolve the conflict by telling `flake8` to ignore what `black`
produces.