tools/link_system_libs: Made it idempotent
Repeated runs of `tools/link_system_libs` used to produce
error messages like

    ln: failed to create symbolic link '/home/user/.local/share/virtualenvs/soundcraft-utils-ABCDEFG/lib/python3.8/site-packages/gi': File exists

Now the script checks if the symlink already exists and
points to the correct place, and will accept that as a
successful script run as well.

If the symlink already exists and points to the wrong place,
an error message will still be produced.