Generate, install and uninstall PNG icons of different sizes
It used to be that `soundcraft_dbus_service --setup` installed
the same 256x256 PNG file as the PNG icon for any number of sizes
which were not 256x256 at all.

We now have all the actual sizes in the git repo so we both
install and uninstall them: The old sizes we installed the wrong
sized icon to (16, 24, 32, 48), a few new sizes which appear to
be in popular use (64, 96, 128, 192), as well as the old 256.

We achieve this by converting the source `*.svg` icon to a number
of `*.png` icon files in different sizes calling `inkscape` from
a script, and then adding those different PNG files to the git tree
alongside the original 256x256 PNG file.  Adding the PNG files to
the git repository avoids a requirement for having `inkscape´
available when installing from source.

All the PNG files each are smaller in bytes than the SVG file is,
and we do not expect (m)any changes to the icon over time, so the
churn in binary files should be acceptable even though we now have
nine PNG files instead of just one.

As icon changes are expected to be very rare, we do not really need
a pre-commit check to run `tools/update-icons.sh`.

This all only makes sense if there are many commonly used desktop
environment installations which cannot properly use icons in the
scalable SVG format and work better when desktop icons and Gtk App
window icons are in PNG format.