setup: Use same src file iteration
Use the same iteration loop over the source files for the D-Bus
files as for the XDG files: Whenever a certain source file comes
up, handle that source file according to its specific needs.

For the `*.service` file, that means a standard string template
substitution and writing the content to a location defined by code.

Other files are just copied over as they are.

Once we use the same loop in multiple places, we can refactor the
code to extract the common code into a separate function.